test again

+++ 0x0C-more_malloc_free/101-mul.c
@@ -9,7 +9,7 @@ void print_error(void);
  * Return: 0
  */
 
-void print_number(int n)
+void print_number(long long int n)
 {
 	if (n < 0)
 	{
@@ -32,7 +32,7 @@ void print_number(int n)
 
 int str_len(char *str)
 {
-	int i = 0;
+	long long int i = 0;
 
 	while (str[i])
 	{
@@ -50,7 +50,7 @@ int str_len(char *str)
 
 int toint(char *str)
 {
-	int i = 0, s = 1, num = 0;
+	long long int i = 0, s = 1, num = 0;
 
 	if (str[i] == '-')
 	{
@@ -97,7 +97,7 @@ void print_error(void)
  */
 int multiply(char *num1_str, char *num2_str)
 {
-	int num1, num2, result;
+	long long int num1, num2, result;
 
 	num1 = toint(num1_str);
 	num2 = toint(num2_str);
@@ -115,7 +115,7 @@ int multiply(char *num1_str, char *num2_str)
 
 int main(int argc, char **argv)
 {
-	int result;
+	long long int result;
 
 	if (argc != 3)
 	{
